Rename endpoint-collection file
